+++ storage-resize-images/CHANGELOG.md
@@ -1,3 +1,11 @@
+## Version 0.2.9
+
+fixed - run an npm audit and bump vulnerable dependencies
+
+fixed - use Promise.allSettled to await various resizes
+
+feature - Optional AI content filtering with Gemini
+
 ## Version 0.2.8
 
 fixed - support '+' character in paths